LayoutTests/inspector/extensions/extensions-panel-expected.txt

CONSOLE MESSAGE: line 8: hello
Tests WebInspector extension API

Started extension.
Running tests...
RUNNING TEST: extension_testCreatePanel
Panel created
{
    createStatusBarButton : <function>
    onHidden : {
        addListener : <function>
        removeListener : <function>
    }
    onSearch : {
        addListener : <function>
        removeListener : <function>
    }
    onShown : {
        addListener : <function>
        removeListener : <function>
    }
    show : <function>
}
Panel shown
Extension panel size correct
RUNNING TEST: extension_testSearch
Panel hidden
Panel searched:
{
    0 : "performSearch"
    1 : "hello"
}
Panel searched:
{
    0 : "cancelSearch"
    1 : undefined
}
RUNNING TEST: extension_testStatusBarButtons
Created a status bar button, dump follows:
{
    onClicked : {
        addListener : <function>
        removeListener : <function>
    }
    update : <function>
}
button1 clicked
Status bar buttons state:
status bar item 0, icon: .../inspector/resources/button1.png, tooltip: 'Button One tooltip', disabled: false
status bar item 1, icon: .../inspector/resources/button2.png, tooltip: 'Button Two tooltip', disabled: true
Updated status bar buttons
Status bar buttons state:
status bar item 0, icon: .../inspector/resources/button1-updated.png, tooltip: 'Button One tooltip', disabled: false
status bar item 1, icon: .../inspector/resources/button2.png, tooltip: 'Button Two updated tooltip', disabled: false
button2 clicked
All tests done.
